requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
pydantic
pydantic-settings
alembic
email-validator
pymysql[rsa]
cryptography
aiofiles
Pillow
python-multipart
bcrypt==3.2.0
python-dateutil